[WINDOWS] Get windows compiling...

+++ Source/core/DataElement.h
@@ -55,9 +55,9 @@ namespace Core {
         inline void Copy(const uint8_t data[], const uint16_t length, const uint16_t offset = 0) {
             ASSERT (offset < _size);
             if (offset < _size) {
-                ASSERT (offset + length <= _size);
+                ASSERT (static_cast<uint32_t>(offset + length) <= _size);
 
-                uint32_t count (offset + length <= _size ? length : _size - offset);
+                uint32_t count (static_cast<uint32_t>(offset + length) <= _size ? length : _size - offset);
 
                 ::memcpy (&(_buffer[offset]), data, count);
             }
@@ -198,11 +198,11 @@ namespace Core {
         template <typename TYPE>
         inline void Align() {
             if (m_Buffer != nullptr) {
-                size_t size (m_MaxSize);
+                size_t size (static_cast<size_t>(m_MaxSize));
                 uint8_t adjust(~0);
                 void* newPointer = m_Buffer;
                 if (std::align(sizeof(TYPE), sizeof(TYPE), newPointer, size)) {
-                    uint8_t adjust (m_MaxSize - size);
+                    uint8_t adjust (static_cast<uint8_t>(m_MaxSize - size));
                     m_Buffer = reinterpret_cast<uint8_t*>(newPointer);
                     m_Size = (adjust < m_Size ? (m_Size - adjust) : 0);
                     m_MaxSize -= adjust;
@@ -228,7 +228,7 @@ namespace Core {
         }
         inline bool operator==(const DataElement& RHS) const
         {
-            return ((m_Size == RHS.m_Size) && (::memcmp(m_Buffer, RHS.m_Buffer, m_Size) == 0));
+            return ((m_Size == RHS.m_Size) && (::memcmp(m_Buffer, RHS.m_Buffer, static_cast<size_t>(m_Size)) == 0));
         }
         inline bool operator!=(const DataElement& RHS) const
         {
